functions exported, cn-deleted
