+++ src/components/ui/infinite-moving-cards.tsx
@@ -1,11 +1,12 @@
 "use client";
 
-import { cn } from "@/src/utils/cn";
+import { cn } from "@/src/utils/classnames";
 import React, { useEffect, useState } from "react";
 import {LinkEnum} from "@/src/utils/route";
 import {Button} from "@/src/components/ui/button";
 import Link from "next/link";
 import {useParams} from "next/navigation";
+import {addAnimation} from "@/src/utils/infinite-moving-card-functions";
 
 export const InfiniteMovingCards = ({
                                         items,
@@ -31,54 +32,11 @@ export const InfiniteMovingCards = ({
 }) => {
     const containerRef = React.useRef<HTMLDivElement>(null);
     const scrollerRef = React.useRef<HTMLUListElement>(null);
-
+    const params = useParams()
     useEffect(() => {
-        addAnimation();
+        addAnimation(speed, direction, containerRef, scrollerRef, setStart);
     }, []);
     const [start, setStart] = useState(false);
-    function addAnimation() {
-        if (containerRef.current && scrollerRef.current) {
-            const scrollerContent = Array.from(scrollerRef.current.children);
-
-            scrollerContent.forEach((item) => {
-                const duplicatedItem = item.cloneNode(true);
-                if (scrollerRef.current) {
-                    scrollerRef.current.appendChild(duplicatedItem);
-                }
-            });
-
-            getDirection();
-            getSpeed();
-            setStart(true);
-        }
-    }
-    const getDirection = () => {
-        if (containerRef.current) {
-            if (direction === "left") {
-                containerRef.current.style.setProperty(
-                    "--animation-direction",
-                    "forwards"
-                );
-            } else {
-                containerRef.current.style.setProperty(
-                    "--animation-direction",
-                    "reverse"
-                );
-            }
-        }
-    };
-    const getSpeed = () => {
-        if (containerRef.current) {
-            if (speed === "fast") {
-                containerRef.current.style.setProperty("--animation-duration", "20s");
-            } else if (speed === "normal") {
-                containerRef.current.style.setProperty("--animation-duration", "40s");
-            } else {
-                containerRef.current.style.setProperty("--animation-duration", "80s");
-            }
-        }
-    };
-    const params = useParams()
     return (
         <div
             ref={containerRef}
